move files utility only used in copilot to copilot domain
