+++ tests/platform_helper/domain/test_copilot.py
@@ -1070,6 +1070,26 @@ def test_is_service_empty_manifest(fakefs):
     )
 
 
+def test_generate_override_files(fakefs):
+    """Test that, given a path to override files and an output directory,
+    generate_override_files copies the required files to the output
+    directory."""
+
+    fakefs.create_file("templates/.gitignore")
+    fakefs.create_file("templates/bin/code.ts")
+    fakefs.create_file("templates/node_modules/package.ts")
+
+    mocks = CopilotMocks()
+    mocks.file_provider = FileProvider
+    Copilot(**mocks.params())._generate_override_files(
+        base_path=Path("."), file_path=Path("templates"), output_dir=Path("output")
+    )
+
+    assert ".gitignore" in os.listdir("/output")
+    assert "code.ts" in os.listdir("/output/bin")
+    assert "node_modules" not in os.listdir("/output")
+
+
 def create_test_manifests(addon_file_contents, fakefs):
     content = yaml.dump({"application": "test-app", "extensions": addon_file_contents})
     fakefs.create_file(PLATFORM_CONFIG_FILE, contents=content)
