+++ tests/platform_helper/utils/test_files.py
@@ -1,26 +1,7 @@
 import os
 from pathlib import Path
 
-from dbt_platform_helper.utils.files import generate_override_files
 from dbt_platform_helper.utils.files import generate_override_files_from_template
-
-
-def test_generate_override_files(fakefs):
-    """Test that, given a path to override files and an output directory,
-    generate_override_files copies the required files to the output
-    directory."""
-
-    fakefs.create_file("templates/.gitignore")
-    fakefs.create_file("templates/bin/code.ts")
-    fakefs.create_file("templates/node_modules/package.ts")
-
-    generate_override_files(
-        base_path=Path("."), file_path=Path("templates"), output_dir=Path("output")
-    )
-
-    assert ".gitignore" in os.listdir("/output")
-    assert "code.ts" in os.listdir("/output/bin")
-    assert "node_modules" not in os.listdir("/output")
 
 
 def test_generate_pipeline_override_files(fakefs):
